Add raise tracking on character sheet #62
