+++ lang/en.json
@@ -329,5 +329,6 @@
   "SVNSEA2E.Small": "Small",
   "SVNSEA2E.IncreaseThreshold": "Increase Threshold by 5",
   "SVNSEA2E.GMIncreasedThreshold": "(GM increased the threshold by 5)",
-  "SVNSEA2E.AddToInitiativeTracker": "Add to initiative tracker"
+  "SVNSEA2E.AddToInitiativeTracker": "Add to initiative tracker",
+  "SVNSEA2E.Initiative": "Raises"
 }
