+++ lang/es.json
@@ -329,5 +329,6 @@
   "SVNSEA2E.Small": "Small",
   "SVNSEA2E.IncreaseThreshold": "Incrementar el umbral en 5",
   "SVNSEA2E.GMIncreasedThreshold": "(el DJ incrementa el umbral en 5)",
-  "SVNSEA2E.AddToInitiativeTracker": "Agregar al rastreador de iniciativas"
+  "SVNSEA2E.AddToInitiativeTracker": "Agregar al rastreador de iniciativas",
+  "SVNSEA2E.Initiative": "Aumentos"
 }
