+++ scss/components/_actors.scss
@@ -16,6 +16,13 @@
   .wounds .pts.bigger {
     padding-left: 37px;
   }
+  
+  .raise-tracking {
+    button {
+      width: 2em;
+      margin: 0 1em;
+    }
+  }
 }
 
 .danger-points {
