+++ scss/components/_sheet.scss
@@ -65,6 +65,11 @@ nav.sheet-tabs {
     max-width: $max-input-width-xs;
     text-align: right;
   }
+
+  .herospts, .wealth {
+    max-width: 30px;
+    margin-right: 10px;
+  }
 }
 
 .tag {
